Move solarlog coordinator to own file
